Moved revocation related stuff to separate directory
