+++ components/connector-service/internal/revocation/mocks/Manager.go
