+++ components/connector-service/internal/revocation/repository.go
@@ -1,4 +1,4 @@
-package revocationlist
+package revocation
 
 import (
 	"k8s.io/api/core/v1"
